tasks-service/pyproject.toml
[project]
name = "tasks-service"
version = "0.1.0"
description = ""
authors = [
    {name = "novoselovsv",email = "novoselovsv@bk.ru"}
]
readme = "README.md"
requires-python = "<4.0,>=3.11"
dependencies = [
    "fastapi[standart] (>=0.121.1,<0.122.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sse-starlette (>=3.0.3,<4.0.0)",
    "faststream[rabbit] (>=0.6.3,<0.7.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "tasks_service", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "isort (>=7.0.0,<8.0.0)"
]
